feat: Testing framework
This commit introduces jest as a testing framework for the LAMP-server. I had to separate out the bootstrap code from the application to make the test harness stop complaining. The raw express application is also required for `supertest` integration testing. I also updated the typescript toolchain to something a bit more modern.
