+++ docker-compose.yaml
@@ -1,0 +1,29 @@
+version: '3'
+services:
+  mongo:
+    image: mongo
+    restart: always
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: root
+      MONGO_INITDB_ROOT_PASSWORD: password
+    ports:
+      - 27017:27017
+  mongo-express:
+    image: mongo-express
+    restart: always
+    ports:
+      - 8081:8081
+    environment:
+      ME_CONFIG_MONGODB_ADMINUSERNAME: root
+      ME_CONFIG_MONGODB_ADMINPASSWORD: password
+  redis:
+    image: redis
+    restart: always
+    ports:
+      - 6379:6379
+  nats:
+    image: nats
+    restart: always
+    ports:
+      - 4222:4222
+      - 8222:8222
