+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "lamp-server",
   "version": "1.0.0",
-  "bin": "build/app.js",
+  "bin": "build/index.js",
   "dependencies": {
     "ajv": "^8.1.0",
     "async-mutex": "^0.2.4",
@@ -22,7 +22,8 @@
     "@types/dockerode": "^2.5.21",
     "@types/express": "^4.17.2",
     "@types/fs-extra": "^5.1.0",
-    "@types/morgan": "^1.7.37",
+    "@types/jest": "^26.0.23",
+    "@types/morgan": "^1.9.2",
     "@types/node": "^13.1.2",
     "@types/node-fetch": "^2.5.5",
     "@types/readable-stream": "^2.3.5",
@@ -36,19 +37,26 @@
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.3",
+    "jest": "^26.6.3",
+    "nodemon": "^2.0.7",
     "pkg": "^4.4.2",
     "prettier": "^2.0.5",
+    "rimraf": "^2.6.3",
     "source-map-support": "^0.5.16",
-    "tsc-watch": "^1.1.39",
+    "ts-jest": "^26.5.5",
+    "ts-node": "^9.1.1",
     "tslint": "^5.20.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-standard": "^8.0.1",
     "typescript": "^3.7.4"
   },
   "scripts": {
-    "start": "tsc-watch --onSuccess \"node -r source-map-support/register ./build/app.js\"",
-    "build": "tsc",
-    "deploy": "pkg . -t node10-alpine-x64,node10-linux-x64,node10-macos-x64 --out-path bin/"
+    "build": "rimraf build && tsc",
+    "coverage": "jest --coverage",
+    "dev": "docker compose up -d && nodemon --exec ts-node ./src/index.ts --watch src",
+    "deploy": "rimraf bin && pkg . -t node10-alpine-x64,node10-linux-x64,node10-macos-x64 --out-path bin/",
+    "start": "npm run build && node build/index.js",
+    "test": "jest --watch"
   },
   "pkg": {
     "scripts": [
@@ -86,5 +94,10 @@
   "prettier": {
     "semi": false,
     "printWidth": 120
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "globalSetup": "./src/test/setup.js"
   }
 }
