+++ src/test/setup.js
@@ -1,0 +1,3 @@
+module.exports = async () => {
+  require('dotenv').config()
+}
